Move create_*_from_config to _config.py and add tests.
