+++ optuna_dashboard/_cli.py
@@ -4,7 +4,6 @@ import argparse
 import os
 from socketserver import ThreadingMixIn
 import sys
-from typing import Any
 from typing import TYPE_CHECKING
 from wsgiref.simple_server import make_server
 from wsgiref.simple_server import WSGIServer
@@ -16,6 +15,8 @@ from optuna.storages import RDBStorage
 
 from . import __version__
 from ._app import create_app
+from ._config import create_artifact_store_from_config
+from ._config import create_llm_provider_from_config
 from ._config import load_config_from_toml
 from ._sql_profiler import register_profiler_view
 from ._storage_url import get_storage
@@ -25,8 +26,6 @@ if TYPE_CHECKING:
     from typing import Literal
 
     from optuna.artifacts._protocol import ArtifactStore
-
-    from .llm.provider import LLMProvider
 
 
 DEBUG = os.environ.get("OPTUNA_DASHBOARD_DEBUG") == "1"
@@ -85,63 +84,6 @@ def auto_select_server(
         return "gunicorn"
     except ImportError:
         return "wsgiref"
-
-
-def create_llm_provider_from_config(config: dict[str, Any]) -> LLMProvider | None:
-    if "llm" not in config:
-        return None
-
-    llm_config = config["llm"]
-
-    if "openai" in llm_config:
-        import optuna_dashboard.llm.openai
-        import openai
-
-        client = openai.OpenAI(**llm_config["openai"]["client"])
-        return optuna_dashboard.llm.openai.OpenAI(
-            client,
-            model=llm_config["openai"]["model"],
-            use_chat_completions_api=llm_config["openai"].get("use_chat_completions_api", False),
-        )
-    elif "azure_openai" in llm_config:
-        import optuna_dashboard.llm.openai
-        import openai
-
-        client = openai.AzureOpenAI(**llm_config["azure_openai"]["client"])
-        return optuna_dashboard.llm.openai.AzureOpenAI(
-            client,
-            model=llm_config["azure_openai"]["model"],
-            use_chat_completions_api=llm_config["azure_openai"].get(
-                "use_chat_completions_api", False
-            ),
-        )
-    else:
-        raise ValueError(
-            "Unsupported LLM provider. Supported providers: 'openai', 'azure_openai'."
-        )
-
-
-def create_artifact_store_from_config(config: dict[str, Any]) -> ArtifactStore | None:
-    if "artifact_store" not in config:
-        return None
-
-    artifact_store: ArtifactStore | None = None
-    if "boto3" in config["artifact_store"]:
-        from optuna.artifacts import Boto3ArtifactStore
-
-        artifact_store = Boto3ArtifactStore(**config["artifact_store"]["boto3"])
-    elif "gcs" in config["artifact_store"]:
-        from optuna.artifacts import GCSArtifactStore
-
-        artifact_store = GCSArtifactStore(**config["artifact_store"]["gcs"])
-    elif "filesystem" in config["artifact_store"]:
-        from optuna.artifacts import FileSystemArtifactStore
-
-        artifact_store = FileSystemArtifactStore(**config["artifact_store"]["filesystem"])
-    else:
-        raise ValueError("Unsupported artifact store configuration.")
-
-    return artifact_store
 
 
 def main() -> None:
